feat: add clear-cache-and-check
This commit adds the above, which purges the cached_htmls and re-runs
all tests, just in case it is wanted.

I also discovered during the above process that some HardwareZone tests
have since become outdated, since the post is deleted.
Hence, a NOT_FOUND check is added to make sure it doesn't keep
requesting the webpage if it does not exist. Furthermore, I have changed
the tests to use a HardwareZone URL that still exists.

To aid in testing, basic logging is also added to see what exactly in
the test has failed.