chore: split Makefile out, use scripts/, use Invoke
The core problem this commit tries to tackle is that the syntax for
linting is terrible. Running [make lint args="file1 file2"] is
especially terrible since autocomplete does not work in quotes like
that.

According to some strangers on the internet, Makefile wasn't meant to be
used this way.

This commit uses Pyinvoke, which is not necessarily a better change,
since it also does not support a nice way to run [make lint]. Hence, the
commit creates a new [scripts/] directory with a bunch of scripts to
start the programs, while developmental commands can live in the Invoke
ecosystem for better organisation. Pyinvoke also allows pretty nice ways
to involve flags and reuse code.

Just, a Makefile replacement which does support such nice multi-variable
extensions for [make lint bla bla2], seems like a good alternative.
However, the syntax is quite cryptic and heavy-handed for such a
lightweight use for now. Hence, this commit does not use Just.