feat(linter): allow program to more elegantly handle malformed JSONs
Upon passing in a malformed JSON in a sea of JSONs, the linter errors
out completely and cryptically, without saying which file caused it.

This commit employs Pydantic to do a type check of the JSON, and if it
is unable to validate it against the type, it is better accumulated and
the main function actually provides a helpful hint of what is going on.